Feature/book crud

+++ src/main/java/uoc/tfg/cvelascofa/pageturner_backend/book/service/GenreServiceImpl.java
@@ -1,12 +1,17 @@
 package uoc.tfg.cvelascofa.pageturner_backend.book.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import uoc.tfg.cvelascofa.pageturner_backend.book.dto.GenreDTO;
 import uoc.tfg.cvelascofa.pageturner_backend.book.entity.Genre;
 import uoc.tfg.cvelascofa.pageturner_backend.book.mapper.GenreMapper;
+import uoc.tfg.cvelascofa.pageturner_backend.book.repository.BookRepository;
 import uoc.tfg.cvelascofa.pageturner_backend.book.repository.GenreRepository;
 import uoc.tfg.cvelascofa.pageturner_backend.book.service.interfaces.GenreService;
+import uoc.tfg.cvelascofa.pageturner_backend.exception.GenreInUseException;
+import uoc.tfg.cvelascofa.pageturner_backend.exception.ResourceNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -20,6 +25,9 @@ public class GenreServiceImpl implements GenreService {
 
     @Autowired
     private GenreMapper genreMapper;
+
+    @Autowired
+    private BookRepository bookRepository;
 
     @Override
     public GenreDTO create(GenreDTO genreDTO) {
@@ -58,7 +66,21 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public void delete(Long id) {
+        Optional<Genre> genre = genreRepository.findById(id);
+        if (genre.isEmpty()) {
+            throw new ResourceNotFoundException("Genre not found");
+        }
+
+        if (bookRepository.existsByGenreId(id)) {
+            throw new GenreInUseException("This genre is still referenced by books");
+        }
         genreRepository.deleteById(id);
+    }
+
+    @Override
+    public Page<GenreDTO> searchGenresPageable(String name, Pageable pageable) {
+        Page<Genre> genresPage = genreRepository.findByNameContainingIgnoreCase(name, pageable);
+        return genresPage.map(genreMapper::toDTO);
     }
 
 }
